+++ tests/lib/Share20/DefaultShareProviderTest.php
@@ -53,9 +53,9 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 	public function setUp() {
 		$this->dbConn = \OC::$server->getDatabaseConnection();
-		$this->userManager = $this->getMock('OCP\IUserManager');
-		$this->groupManager = $this->getMock('OCP\IGroupManager');
-		$this->rootFolder = $this->getMock('OCP\Files\IRootFolder');
+		$this->userManager = $this->getMockBuilder('OCP\IUserManager')->getMock();
+		$this->groupManager = $this->getMockBuilder('OCP\IGroupManager')->getMock();
+		$this->rootFolder = $this->getMockBuilder('OCP\Files\IRootFolder')->getMock();
 
 		$this->userManager->expects($this->any())->method('userExists')->willReturn(true);
 
@@ -136,13 +136,13 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$id = $qb->getLastInsertId();
 
-		$sharedBy = $this->getMock('OCP\IUser');
+		$sharedBy = $this->getMockBuilder('OCP\IUser')->getMock();
 		$sharedBy->method('getUID')->willReturn('sharedBy');
-		$shareOwner = $this->getMock('OCP\IUser');
+		$shareOwner = $this->getMockBuilder('OCP\IUser')->getMock();
 		$shareOwner->method('getUID')->willReturn('shareOwner');
 
-		$ownerPath = $this->getMock('\OCP\Files\File');
-		$shareOwnerFolder = $this->getMock('\OCP\Files\Folder');
+		$ownerPath = $this->getMockBuilder('OCP\Files\File')->getMock();
+		$shareOwnerFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$shareOwnerFolder->method('getById')->with(42)->willReturn([$ownerPath]);
 
 		$this->rootFolder
@@ -218,9 +218,9 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$id = $qb->getLastInsertId();
 
-		$ownerPath = $this->getMock('\OCP\Files\File');
+		$ownerPath = $this->getMockBuilder('OCP\Files\File')->getMock();
 
-		$shareOwnerFolder = $this->getMock('\OCP\Files\Folder');
+		$shareOwnerFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$shareOwnerFolder->method('getById')->with(42)->willReturn([$ownerPath]);
 
 		$this->rootFolder
@@ -263,8 +263,8 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		// Get the id
 		$id = $qb->getLastInsertId();
 
-		$ownerPath = $this->getMock('\OCP\Files\Folder');
-		$shareOwnerFolder = $this->getMock('\OCP\Files\Folder');
+		$ownerPath = $this->getMockBuilder('OCP\Files\Folder')->getMock();
+		$shareOwnerFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$shareOwnerFolder->method('getById')->with(42)->willReturn([$ownerPath]);
 
 		$this->rootFolder
@@ -291,15 +291,15 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_GROUP, 'group0', 'user0', 'user0', 'file', 42, 'myTarget', 31, null, null);
 		$this->addShareToDB(2, 'user1', 'user0', 'user0', 'file', 42, 'userTarget', 0, null, null, $id);
 
-		$user0 = $this->getMock('OCP\IUser');
+		$user0 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user0->method('getUID')->willReturn('user0');
-		$user1 = $this->getMock('OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
 
-		$group0 = $this->getMock('OCP\IGroup');
+		$group0 = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group0->method('inGroup')->with($user1)->willReturn(true);
 
-		$node = $this->getMock('\OCP\Files\Folder');
+		$node = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$node->method('getId')->willReturn(42);
 
 		$this->rootFolder->method('getUserFolder')->with('user0')->will($this->returnSelf());
@@ -325,50 +325,6 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertEquals('userTarget', $share->getTarget());
 	}
 
-	public function testGetShareByIdLinkShare() {
-		$qb = $this->dbConn->getQueryBuilder();
-
-		$qb->insert('share')
-			->values([
-				'share_type' => $qb->expr()->literal(\OCP\Share::SHARE_TYPE_LINK),
-				'share_with' => $qb->expr()->literal('sharedWith'),
-				'uid_owner' => $qb->expr()->literal('shareOwner'),
-				'uid_initiator' => $qb->expr()->literal('sharedBy'),
-				'item_type'   => $qb->expr()->literal('file'),
-				'file_source' => $qb->expr()->literal(42),
-				'file_target' => $qb->expr()->literal('myTarget'),
-				'permissions' => $qb->expr()->literal(13),
-				'token' => $qb->expr()->literal('token'),
-				'expiration' => $qb->expr()->literal('2000-01-02 00:00:00'),
-			]);
-		$this->assertEquals(1, $qb->execute());
-
-		$id = $qb->getLastInsertId();
-
-		$ownerPath = $this->getMock('\OCP\Files\Folder');
-		$shareOwnerFolder = $this->getMock('\OCP\Files\Folder');
-		$shareOwnerFolder->method('getById')->with(42)->willReturn([$ownerPath]);
-
-		$this->rootFolder
-				->method('getUserFolder')
-				->will($this->returnValueMap([
-						['shareOwner', $shareOwnerFolder],
-				]));
-
-		$share = $this->provider->getShareById($id);
-
-		$this->assertEquals($id, $share->getId());
-		$this->assertEquals(\OCP\Share::SHARE_TYPE_LINK, $share->getShareType());
-		$this->assertEquals('sharedWith', $share->getPassword());
-		$this->assertEquals('sharedBy', $share->getSharedBy());
-		$this->assertEquals('shareOwner', $share->getShareOwner());
-		$this->assertEquals($ownerPath, $share->getNode());
-		$this->assertEquals(13, $share->getPermissions());
-		$this->assertEquals('token', $share->getToken());
-		$this->assertEquals(\DateTime::createFromFormat('Y-m-d H:i:s', '2000-01-02 00:00:00'), $share->getExpirationDate());
-		$this->assertEquals('myTarget', $share->getTarget());
-	}
-
 	public function testDeleteSingleShare() {
 		$qb = $this->dbConn->getQueryBuilder();
 		$qb->insert('share')
@@ -385,7 +341,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$id = $qb->getLastInsertId();
 
-		$share = $this->getMock('OCP\Share\IShare');
+		$share = $this->getMockBuilder('OCP\Share\IShare')->getMock();
 		$share->method('getId')->willReturn($id);
 
 		$provider = $this->getMockBuilder('OC\Share20\DefaultShareProvider')
@@ -473,7 +429,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			]);
 		$this->assertEquals(1, $qb->execute());
 
-		$share = $this->getMock('OCP\Share\IShare');
+		$share = $this->getMockBuilder('OCP\Share\IShare')->getMock();
 		$share->method('getId')->willReturn($id);
 		$share->method('getShareType')->willReturn(\OCP\Share::SHARE_TYPE_GROUP);
 
@@ -548,8 +504,8 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			]);
 		$qb->execute();
 
-		$ownerPath = $this->getMock('\OCP\Files\Folder');
-		$ownerFolder = $this->getMock('\OCP\Files\Folder');
+		$ownerPath = $this->getMockBuilder('\OCP\Files\Folder')->getMock();
+		$ownerFolder = $this->getMockBuilder('\OCP\Files\Folder')->getMock();
 		$ownerFolder->method('getById')->willReturn([$ownerPath]);
 
 		$this->rootFolder
@@ -558,7 +514,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 				['shareOwner', $ownerFolder],
 			]));
 
-		$share = $this->getMock('\OCP\Share\IShare');
+		$share = $this->getMockBuilder('\OCP\Share\IShare')->getMock();
 		$share->method('getId')->willReturn($id);
 
 		$children = $this->provider->getChildren($share);
@@ -591,15 +547,15 @@ class DefaultShareProviderTest extends \Test\TestCase {
 	public function testCreateUserShare() {
 		$share = new \OC\Share20\Share($this->rootFolder, $this->userManager);
 
-		$shareOwner = $this->getMock('OCP\IUser');
+		$shareOwner = $this->getMockBuilder('OCP\IUser')->getMock();
 		$shareOwner->method('getUID')->WillReturn('shareOwner');
 
-		$path = $this->getMock('\OCP\Files\File');
+		$path = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$path->method('getId')->willReturn(100);
 		$path->method('getOwner')->willReturn($shareOwner);
 
-		$ownerFolder = $this->getMock('OCP\Files\Folder');
-		$userFolder = $this->getMock('OCP\Files\Folder');
+		$ownerFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
+		$userFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$this->rootFolder
 			->method('getUserFolder')
 			->will($this->returnValueMap([
@@ -639,15 +595,15 @@ class DefaultShareProviderTest extends \Test\TestCase {
 	public function testCreateGroupShare() {
 		$share = new \OC\Share20\Share($this->rootFolder, $this->userManager);
 
-		$shareOwner = $this->getMock('\OCP\IUser');
+		$shareOwner = $this->getMockBuilder('OCP\IUser')->getMock();
 		$shareOwner->method('getUID')->willReturn('shareOwner');
 
-		$path = $this->getMock('\OCP\Files\Folder');
+		$path = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$path->method('getId')->willReturn(100);
 		$path->method('getOwner')->willReturn($shareOwner);
 
-		$ownerFolder = $this->getMock('OCP\Files\Folder');
-		$userFolder = $this->getMock('OCP\Files\Folder');
+		$ownerFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
+		$userFolder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$this->rootFolder
 			->method('getUserFolder')
 			->will($this->returnValueMap([
@@ -682,91 +638,6 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertSame('/target', $share2->getTarget());
 		$this->assertLessThanOrEqual(new \DateTime(), $share2->getShareTime());
 		$this->assertSame($path, $share2->getNode());
-	}
-
-	public function testCreateLinkShare() {
-		$share = new \OC\Share20\Share($this->rootFolder, $this->userManager);
-
-		$shareOwner = $this->getMock('\OCP\IUser');
-		$shareOwner->method('getUID')->willReturn('shareOwner');
-
-		$path = $this->getMock('\OCP\Files\Folder');
-		$path->method('getId')->willReturn(100);
-		$path->method('getOwner')->willReturn($shareOwner);
-
-		$ownerFolder = $this->getMock('OCP\Files\Folder');
-		$userFolder = $this->getMock('OCP\Files\Folder');
-		$this->rootFolder
-				->method('getUserFolder')
-				->will($this->returnValueMap([
-						['sharedBy', $userFolder],
-						['shareOwner', $ownerFolder],
-				]));
-
-		$userFolder->method('getById')
-				->with(100)
-				->willReturn([$path]);
-		$ownerFolder->method('getById')
-				->with(100)
-				->willReturn([$path]);
-
-		$share->setShareType(\OCP\Share::SHARE_TYPE_LINK);
-		$share->setSharedBy('sharedBy');
-		$share->setShareOwner('shareOwner');
-		$share->setNode($path);
-		$share->setPermissions(1);
-		$share->setPassword('password');
-		$share->setToken('token');
-		$expireDate = new \DateTime();
-		$share->setExpirationDate($expireDate);
-		$share->setTarget('/target');
-
-		$share2 = $this->provider->create($share);
-
-		$this->assertNotNull($share2->getId());
-		$this->assertSame('ocinternal:'.$share2->getId(), $share2->getFullId());
-		$this->assertSame(\OCP\Share::SHARE_TYPE_LINK, $share2->getShareType());
-		$this->assertSame('sharedBy', $share2->getSharedBy());
-		$this->assertSame('shareOwner', $share2->getShareOwner());
-		$this->assertSame(1, $share2->getPermissions());
-		$this->assertSame('/target', $share2->getTarget());
-		$this->assertLessThanOrEqual(new \DateTime(), $share2->getShareTime());
-		$this->assertSame($path, $share2->getNode());
-		$this->assertSame('password', $share2->getPassword());
-		$this->assertSame('token', $share2->getToken());
-		$this->assertEquals($expireDate, $share2->getExpirationDate());
-	}
-
-	public function testGetShareByToken() {
-		$qb = $this->dbConn->getQueryBuilder();
-
-		$qb->insert('share')
-			->values([
-				'share_type'    => $qb->expr()->literal(\OCP\Share::SHARE_TYPE_LINK),
-				'share_with'    => $qb->expr()->literal('password'),
-				'uid_owner'     => $qb->expr()->literal('shareOwner'),
-				'uid_initiator' => $qb->expr()->literal('sharedBy'),
-				'item_type'     => $qb->expr()->literal('file'),
-				'file_source'   => $qb->expr()->literal(42),
-				'file_target'   => $qb->expr()->literal('myTarget'),
-				'permissions'   => $qb->expr()->literal(13),
-				'token'         => $qb->expr()->literal('secrettoken'),
-			]);
-		$qb->execute();
-		$id = $qb->getLastInsertId();
-
-		$file = $this->getMock('\OCP\Files\File');
-
-		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
-		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
-
-		$share = $this->provider->getShareByToken('secrettoken');
-		$this->assertEquals($id, $share->getId());
-		$this->assertSame('shareOwner', $share->getShareOwner());
-		$this->assertSame('sharedBy', $share->getSharedBy());
-		$this->assertSame('secrettoken', $share->getToken());
-		$this->assertSame('password', $share->getPassword());
-		$this->assertSame(null, $share->getSharedWith());
 	}
 
 	/**
@@ -806,7 +677,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			]);
 		$this->assertEquals(1, $qb->execute());
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
 
@@ -853,20 +724,20 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$groups = [];
 		foreach(range(0, 100) as $i) {
-			$group = $this->getMock('\OCP\IGroup');
+			$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 			$group->method('getGID')->willReturn('group'.$i);
 			$groups[] = $group;
 		}
 
-		$group = $this->getMock('\OCP\IGroup');
+		$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group->method('getGID')->willReturn('sharedWith');
 		$groups[] = $group;
 
-		$user = $this->getMock('\OCP\IUser');
+		$user = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user->method('getUID')->willReturn('sharedWith');
-		$owner = $this->getMock('\OCP\IUser');
+		$owner = $this->getMockBuilder('OCP\IUser')->getMock();
 		$owner->method('getUID')->willReturn('shareOwner');
-		$initiator = $this->getMock('\OCP\IUser');
+		$initiator = $this->getMockBuilder('OCP\IUser')->getMock();
 		$initiator->method('getUID')->willReturn('sharedBy');
 
 		$this->userManager->method('get')->willReturnMap([
@@ -877,7 +748,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->groupManager->method('getUserGroups')->with($user)->willReturn($groups);
 		$this->groupManager->method('get')->with('sharedWith')->willReturn($group);
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
 
@@ -944,15 +815,15 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			]);
 		$this->assertEquals(1, $qb->execute());
 
-		$group = $this->getMock('\OCP\IGroup');
+		$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group->method('getGID')->willReturn('sharedWith');
 		$groups = [$group];
 
-		$user = $this->getMock('\OCP\IUser');
+		$user = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user->method('getUID')->willReturn('user');
-		$owner = $this->getMock('\OCP\IUser');
+		$owner = $this->getMockBuilder('OCP\IUser')->getMock();
 		$owner->method('getUID')->willReturn('shareOwner');
-		$initiator = $this->getMock('\OCP\IUser');
+		$initiator = $this->getMockBuilder('OCP\IUser')->getMock();
 		$initiator->method('getUID')->willReturn('sharedBy');
 
 		$this->userManager->method('get')->willReturnMap([
@@ -963,7 +834,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->groupManager->method('getUserGroups')->with($user)->willReturn($groups);
 		$this->groupManager->method('get')->with('sharedWith')->willReturn($group);
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
 
@@ -986,9 +857,9 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_USER, 'user0', 'user1', 'user1',
 			'file', 43, 'myTarget', 31, null, null, null);
 
-		$user0 = $this->getMock('\OCP\IUser');
+		$user0 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user0->method('getUID')->willReturn('user0');
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
 
 		$this->userManager->method('get')->willReturnMap([
@@ -996,7 +867,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			['user1', $user1],
 		]);
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(43);
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(43)->willReturn([$file]);
@@ -1019,9 +890,9 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_GROUP, 'group0', 'user1', 'user1',
 			'file', 43, 'myTarget', 31, null, null, null);
 
-		$user0 = $this->getMock('\OCP\IUser');
+		$user0 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user0->method('getUID')->willReturn('user0');
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
 
 		$this->userManager->method('get')->willReturnMap([
@@ -1029,13 +900,13 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			['user1', $user1],
 		]);
 
-		$group0 = $this->getMock('\OCP\IGroup');
+		$group0 = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group0->method('getGID')->willReturn('group0');
 
 		$this->groupManager->method('get')->with('group0')->willReturn($group0);
 		$this->groupManager->method('getUserGroups')->with($user0)->willReturn([$group0]);
 
-		$node = $this->getMock('\OCP\Files\Folder');
+		$node = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$node->method('getId')->willReturn(43);
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(43)->willReturn([$node]);
@@ -1083,7 +954,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			]);
 		$this->assertEquals(1, $qb->execute());
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
 
@@ -1131,7 +1002,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			]);
 		$this->assertEquals(1, $qb->execute());
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(42);
 		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
@@ -1180,7 +1051,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertEquals(1, $qb->execute());
 		$id2 = $qb->getLastInsertId();
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(42);
 		$this->rootFolder->method('getUserFolder')->with('shareOwner')->will($this->returnSelf());
 		$this->rootFolder->method('getById')->with(42)->willReturn([$file]);
@@ -1223,21 +1094,21 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertEquals(1, $stmt);
 		$id = $qb->getLastInsertId();
 
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
-		$user2 = $this->getMock('\OCP\IUser');
+		$user2 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user2->method('getUID')->willReturn('user2');
 		$this->userManager->method('get')->will($this->returnValueMap([
 			['user1', $user1],
 			['user2', $user2],
 		]));
 
-		$group = $this->getMock('\OCP\IGroup');
+		$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group->method('getGID')->willReturn('group');
 		$group->method('inGroup')->with($user2)->willReturn(true);
 		$this->groupManager->method('get')->with('group')->willReturn($group);
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(1);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1294,21 +1165,21 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			])->execute();
 		$this->assertEquals(1, $stmt);
 
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
-		$user2 = $this->getMock('\OCP\IUser');
+		$user2 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user2->method('getUID')->willReturn('user2');
 		$this->userManager->method('get')->will($this->returnValueMap([
 			['user1', $user1],
 			['user2', $user2],
 		]));
 
-		$group = $this->getMock('\OCP\IGroup');
+		$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group->method('getGID')->willReturn('group');
 		$group->method('inGroup')->with($user2)->willReturn(true);
 		$this->groupManager->method('get')->with('group')->willReturn($group);
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(1);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1354,21 +1225,21 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertEquals(1, $stmt);
 		$id = $qb->getLastInsertId();
 
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
-		$user2 = $this->getMock('\OCP\IUser');
+		$user2 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user2->method('getUID')->willReturn('user2');
 		$this->userManager->method('get')->will($this->returnValueMap([
 			['user1', $user1],
 			['user2', $user2],
 		]));
 
-		$group = $this->getMock('\OCP\IGroup');
+		$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group->method('getGID')->willReturn('group');
 		$group->method('inGroup')->with($user2)->willReturn(false);
 		$this->groupManager->method('get')->with('group')->willReturn($group);
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(1);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1395,16 +1266,16 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertEquals(1, $stmt);
 		$id = $qb->getLastInsertId();
 
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
-		$user2 = $this->getMock('\OCP\IUser');
+		$user2 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user2->method('getUID')->willReturn('user2');
 		$this->userManager->method('get')->will($this->returnValueMap([
 			['user1', $user1],
 			['user2', $user2],
 		]));
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(1);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1446,17 +1317,17 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertEquals(1, $stmt);
 		$id = $qb->getLastInsertId();
 
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
-		$user2 = $this->getMock('\OCP\IUser');
+		$user2 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user2->method('getUID')->willReturn('user2');
-		$user3 = $this->getMock('\OCP\IUser');
+		$user3 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$this->userManager->method('get')->will($this->returnValueMap([
 			['user1', $user1],
 			['user2', $user2],
 		]));
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(1);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1467,50 +1338,13 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->provider->deleteFromSelf($share, $user3);
 	}
 
-	/**
-	 * @expectedException \OC\Share20\Exception\ProviderException
-	 * @expectedExceptionMessage Invalid shareType
-	 */
-	public function testDeleteFromSelfLink() {
-		$qb = $this->dbConn->getQueryBuilder();
-		$stmt = $qb->insert('share')
-			->values([
-				'share_type' => $qb->expr()->literal(\OCP\Share::SHARE_TYPE_LINK),
-				'uid_owner' => $qb->expr()->literal('user1'),
-				'uid_initiator' => $qb->expr()->literal('user1'),
-				'item_type' => $qb->expr()->literal('file'),
-				'file_source' => $qb->expr()->literal(1),
-				'file_target' => $qb->expr()->literal('myTarget1'),
-				'permissions' => $qb->expr()->literal(2),
-				'token' => $qb->expr()->literal('token'),
-			])->execute();
-		$this->assertEquals(1, $stmt);
-		$id = $qb->getLastInsertId();
-
-		$user1 = $this->getMock('\OCP\IUser');
-		$user1->method('getUID')->willReturn('user1');
-		$this->userManager->method('get')->will($this->returnValueMap([
-			['user1', $user1],
-		]));
-
-		$file = $this->getMock('\OCP\Files\File');
-		$file->method('getId')->willReturn(1);
-
-		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
-		$this->rootFolder->method('getById')->with(1)->willReturn([$file]);
-
-		$share = $this->provider->getShareById($id);
-
-		$this->provider->deleteFromSelf($share, $user1);
-	}
-
 	public function testUpdateUser() {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_USER, 'user0', 'user1', 'user2',
 			'file', 42, 'target', 31, null, null);
 
 		$users = [];
 		for($i = 0; $i < 6; $i++) {
-			$user = $this->getMock('\OCP\IUser');
+			$user = $this->getMockBuilder('OCP\IUser')->getMock();
 			$user->method('getUID')->willReturn('user'.$i);
 			$users['user'.$i] = $user;
 		}
@@ -1521,14 +1355,14 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			})
 		);
 
-		$file1 = $this->getMock('\OCP\Files\File');
+		$file1 = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file1->method('getId')->willReturn(42);
-		$file2 = $this->getMock('\OCP\Files\File');
+		$file2 = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file2->method('getId')->willReturn(43);
 
-		$folder1 = $this->getMock('\OCP\Files\Folder');
+		$folder1 = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder1->method('getById')->with(42)->willReturn([$file1]);
-		$folder2 = $this->getMock('\OCP\Files\Folder');
+		$folder2 = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder2->method('getById')->with(43)->willReturn([$file2]);
 
 		$this->rootFolder->method('getUserFolder')->will($this->returnValueMap([
@@ -1553,111 +1387,13 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$this->assertSame(1, $share2->getPermissions());
 	}
 
-	public function testUpdateLink() {
-		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_LINK, null, 'user1', 'user2',
-			'file', 42, 'target', 31, null, null);
-
-		$users = [];
-		for($i = 0; $i < 6; $i++) {
-			$user = $this->getMock('\OCP\IUser');
-			$user->method('getUID')->willReturn('user'.$i);
-			$users['user'.$i] = $user;
-		}
-
-		$this->userManager->method('get')->will(
-			$this->returnCallback(function($userId) use ($users) {
-				return $users[$userId];
-			})
-		);
-
-		$file1 = $this->getMock('\OCP\Files\File');
-		$file1->method('getId')->willReturn(42);
-		$file2 = $this->getMock('\OCP\Files\File');
-		$file2->method('getId')->willReturn(43);
-
-		$folder1 = $this->getMock('\OCP\Files\Folder');
-		$folder1->method('getById')->with(42)->willReturn([$file1]);
-		$folder2 = $this->getMock('\OCP\Files\Folder');
-		$folder2->method('getById')->with(43)->willReturn([$file2]);
-
-		$this->rootFolder->method('getUserFolder')->will($this->returnValueMap([
-			['user2', $folder1],
-			['user5', $folder2],
-		]));
-
-		$share = $this->provider->getShareById($id);
-
-		$share->setPassword('password');
-		$share->setSharedBy('user4');
-		$share->setShareOwner('user5');
-		$share->setNode($file2);
-		$share->setPermissions(1);
-
-		$share2 = $this->provider->update($share);
-
-		$this->assertEquals($id, $share2->getId());
-		$this->assertEquals('password', $share->getPassword());
-		$this->assertSame('user4', $share2->getSharedBy());
-		$this->assertSame('user5', $share2->getShareOwner());
-		$this->assertSame(1, $share2->getPermissions());
-	}
-
-	public function testUpdateLinkRemovePassword() {
-		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_LINK, 'foo', 'user1', 'user2',
-			'file', 42, 'target', 31, null, null);
-
-		$users = [];
-		for($i = 0; $i < 6; $i++) {
-			$user = $this->getMock('\OCP\IUser');
-			$user->method('getUID')->willReturn('user'.$i);
-			$users['user'.$i] = $user;
-		}
-
-		$this->userManager->method('get')->will(
-			$this->returnCallback(function($userId) use ($users) {
-				return $users[$userId];
-			})
-		);
-
-		$file1 = $this->getMock('\OCP\Files\File');
-		$file1->method('getId')->willReturn(42);
-		$file2 = $this->getMock('\OCP\Files\File');
-		$file2->method('getId')->willReturn(43);
-
-		$folder1 = $this->getMock('\OCP\Files\Folder');
-		$folder1->method('getById')->with(42)->willReturn([$file1]);
-		$folder2 = $this->getMock('\OCP\Files\Folder');
-		$folder2->method('getById')->with(43)->willReturn([$file2]);
-
-		$this->rootFolder->method('getUserFolder')->will($this->returnValueMap([
-			['user2', $folder1],
-			['user5', $folder2],
-		]));
-
-		$share = $this->provider->getShareById($id);
-
-		$share->setPassword(null);
-		$share->setSharedBy('user4');
-		$share->setShareOwner('user5');
-		$share->setNode($file2);
-		$share->setPermissions(1);
-
-		$share2 = $this->provider->update($share);
-
-		$this->assertEquals($id, $share2->getId());
-		$this->assertEquals(null, $share->getPassword());
-		$this->assertSame('user4', $share2->getSharedBy());
-		$this->assertSame('user5', $share2->getShareOwner());
-		$this->assertSame(1, $share2->getPermissions());
-	}
-
 	public function testUpdateGroupNoSub() {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_GROUP, 'group0', 'user1', 'user2',
 			'file', 42, 'target', 31, null, null);
 
 		$users = [];
 		for($i = 0; $i < 6; $i++) {
-			$user = $this->getMock('\OCP\IUser');
+			$user = $this->getMockBuilder('OCP\IUser')->getMock();
 			$user->method('getUID')->willReturn('user'.$i);
 			$users['user'.$i] = $user;
 		}
@@ -1670,7 +1406,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$groups = [];
 		for($i = 0; $i < 2; $i++) {
-			$group = $this->getMock('\OCP\IGroup');
+			$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 			$group->method('getGID')->willReturn('group'.$i);
 			$groups['group'.$i] = $group;
 		}
@@ -1681,14 +1417,14 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			})
 		);
 
-		$file1 = $this->getMock('\OCP\Files\File');
+		$file1 = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file1->method('getId')->willReturn(42);
-		$file2 = $this->getMock('\OCP\Files\File');
+		$file2 = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file2->method('getId')->willReturn(43);
 
-		$folder1 = $this->getMock('\OCP\Files\Folder');
+		$folder1 = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder1->method('getById')->with(42)->willReturn([$file1]);
-		$folder2 = $this->getMock('\OCP\Files\Folder');
+		$folder2 = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder2->method('getById')->with(43)->willReturn([$file2]);
 
 		$this->rootFolder->method('getUserFolder')->will($this->returnValueMap([
@@ -1726,7 +1462,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$users = [];
 		for($i = 0; $i < 6; $i++) {
-			$user = $this->getMock('\OCP\IUser');
+			$user = $this->getMockBuilder('OCP\IUser')->getMock();
 			$user->method('getUID')->willReturn('user'.$i);
 			$users['user'.$i] = $user;
 		}
@@ -1739,7 +1475,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 
 		$groups = [];
 		for($i = 0; $i < 2; $i++) {
-			$group = $this->getMock('\OCP\IGroup');
+			$group = $this->getMockBuilder('OCP\IGroup')->getMock();
 			$group->method('getGID')->willReturn('group'.$i);
 			$groups['group'.$i] = $group;
 		}
@@ -1750,14 +1486,14 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			})
 		);
 
-		$file1 = $this->getMock('\OCP\Files\File');
+		$file1 = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file1->method('getId')->willReturn(42);
-		$file2 = $this->getMock('\OCP\Files\File');
+		$file2 = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file2->method('getId')->willReturn(43);
 
-		$folder1 = $this->getMock('\OCP\Files\Folder');
+		$folder1 = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder1->method('getById')->with(42)->willReturn([$file1]);
-		$folder2 = $this->getMock('\OCP\Files\Folder');
+		$folder2 = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder2->method('getById')->with(43)->willReturn([$file2]);
 
 		$this->rootFolder->method('getUserFolder')->will($this->returnValueMap([
@@ -1809,9 +1545,9 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_USER, 'user0', 'user1', 'user1', 'file',
 			42, 'mytaret', 31, null, null);
 
-		$user0 = $this->getMock('\OCP\IUser');
+		$user0 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user0->method('getUID')->willReturn('user0');
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
 
 		$this->userManager->method('get')->will($this->returnValueMap([
@@ -1819,7 +1555,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			['user1', $user1],
 		]));
 
-		$file = $this->getMock('\OCP\Files\File');
+		$file = $this->getMockBuilder('OCP\Files\File')->getMock();
 		$file->method('getId')->willReturn(42);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1838,12 +1574,12 @@ class DefaultShareProviderTest extends \Test\TestCase {
 		$id = $this->addShareToDB(\OCP\Share::SHARE_TYPE_GROUP, 'group0', 'user1', 'user1', 'file',
 			42, 'mytaret', 31, null, null);
 
-		$user0 = $this->getMock('\OCP\IUser');
+		$user0 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user0->method('getUID')->willReturn('user0');
-		$user1 = $this->getMock('\OCP\IUser');
+		$user1 = $this->getMockBuilder('OCP\IUser')->getMock();
 		$user1->method('getUID')->willReturn('user1');
 
-		$group0 = $this->getMock('\OCP\IGroup');
+		$group0 = $this->getMockBuilder('OCP\IGroup')->getMock();
 		$group0->method('getGID')->willReturn('group0');
 		$group0->method('inGroup')->with($user0)->willReturn(true);
 
@@ -1854,7 +1590,7 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			['user1', $user1],
 		]));
 
-		$folder = $this->getMock('\OCP\Files\Folder');
+		$folder = $this->getMockBuilder('OCP\Files\Folder')->getMock();
 		$folder->method('getId')->willReturn(42);
 
 		$this->rootFolder->method('getUserFolder')->with('user1')->will($this->returnSelf());
@@ -1886,9 +1622,9 @@ class DefaultShareProviderTest extends \Test\TestCase {
 			// The group c is still valid but user c is deleted so group share stays
 			[\OCP\Share::SHARE_TYPE_GROUP, 'a', 'b', 'c', 'c', false],
 			[\OCP\Share::SHARE_TYPE_GROUP, 'a', 'b', 'c', 'd', false],
-			[\OCP\Share::SHARE_TYPE_LINK, 'a', 'b', 'c', 'a', true],
+			[\OCP\Share::SHARE_TYPE_LINK, 'a', 'b', 'c', 'a', false],
 			// To avoid invisible link shares delete initiated link shares as well (see #22327)
-			[\OCP\Share::SHARE_TYPE_LINK, 'a', 'b', 'c', 'b', true],
+			[\OCP\Share::SHARE_TYPE_LINK, 'a', 'b', 'c', 'b', false],
 			[\OCP\Share::SHARE_TYPE_LINK, 'a', 'b', 'c', 'c', false],
 			[\OCP\Share::SHARE_TYPE_LINK, 'a', 'b', 'c', 'd', false],
 		];
